lnrpc: add GetPeer api
Add an api method to get one specific peer. This call is similar to ListPeers,
except it only returns a single peer, selected by pubkey, if that peer is
connected at the time. This can be useful if you want to know whether a specific
peer is active at the current moment.